server/user/requirements.txt
apistar==0.3.9
PyMongo==3.4.0
python-bsonjs==0.2.0